+++ src/pingus/components/smallmap.cpp
@@ -57,7 +57,7 @@ SmallMap::draw(DrawingContext& gc)
 
   if (world->get_width() > gc.get_width())
   {
-    int rwidth = int(gc.get_width()  * rect.get_width()  / world->get_width());
+    int rwidth = gc.get_width()  * rect.get_width()  / world->get_width();
     view_rect.left  = rect.left + (of.x * rect.get_width()  / world->get_width()) - rwidth/2;
     view_rect.right = view_rect.left + rwidth;
   }
@@ -69,7 +69,7 @@ SmallMap::draw(DrawingContext& gc)
 
   if (world->get_height() > gc.get_height())
   {
-    int rheight = int(gc.get_height() * rect.get_height() / world->get_height());
+    int rheight = gc.get_height() * rect.get_height() / world->get_height();
     view_rect.top    = rect.top + (of.y * rect.get_height() / world->get_height()) - rheight/2;
     view_rect.bottom = view_rect.top + rheight;
   }
@@ -118,8 +118,8 @@ SmallMap::draw_sprite(const Sprite& sprite, const Vector3f& pos)
 bool
 SmallMap::is_at (int x, int y)
 {
-  return (x > rect.left && x < rect.left + static_cast<int>(rect.get_width())
-          && y > rect.top && y < rect.top + static_cast<int>(rect.get_height()));
+  return (x > rect.left && x < rect.left + rect.get_width()
+          && y > rect.top && y < rect.top + rect.get_height());
 }
 
 void
@@ -130,8 +130,8 @@ SmallMap::on_pointer_move (int x, int y)
 
   if (scroll_mode)
   {
-    cx = (x - rect.left) * static_cast<int>(world->get_width()  / rect.get_width());
-    cy = (y - rect.top) * static_cast<int>(world->get_height() / rect.get_height());
+    cx = (x - rect.left) * (world->get_width()  / rect.get_width());
+    cy = (y - rect.top) * (world->get_height() / rect.get_height());
 
     playfield->set_viewpoint(cx, cy);
   }
@@ -145,8 +145,8 @@ SmallMap::on_primary_button_press (int x, int y)
   // set view to the given COs
   int cx, cy;
   World* world = server->get_world();
-  cx = (x - rect.left) * int(world->get_width()) / rect.get_width();
-  cy = (y - rect.top) * int(world->get_height()) / rect.get_height();
+  cx = (x - rect.left) * world->get_width() / rect.get_width();
+  cy = (y - rect.top) * world->get_height() / rect.get_height();
   playfield->set_viewpoint(cx, cy);
 }
